+++ CHANGELOG.md
@@ -19,6 +19,9 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 
 - xWebAdministration
   - Update GitVersion.yml with the correct regular expression.
+- xWebsite
+  - Fixed HTTPS binding issue causing failure when CertificateSubject matches
+    multiple certificates.
 
 ## [3.1.0] - 2019-12-30